Improve date parsing to be engine independent
Safari/Mobile Safari isn't able to parse ISO-8601 date formats
using the default date constructor.
Make sure the parsing of dates is the same in all engines
using dtox by using moment.js for it.

This change introduces a format parameter which can be specified
in the field options and then parses the date string using the strict
settings of moment.